fix: changing the priority of checking a cell to checking if it is a ghost and resetting each cell after an action is performed on it

+++ src/model/Ghost.java
@@ -160,11 +160,11 @@ public class Ghost {
             positionY += moveSpeed * directionY / length;
         }
     }
-    public int getX() {
+    public float getX() {
         return Math.round(positionX);
     }
 
-    public int getY() {
+    public float getY() {
         return Math.round(positionY);
     }
